add: validationsHelper e refatorando alguns pontos
