+++ back-end/controllers/saleController.js
@@ -3,7 +3,6 @@ const { saleServices } = require('../services');
 
 const createSale = async (request, response) => {
   try {
-    console.log('entrou no controller')
     const { email, total_price, delivery_address, delivery_number, products_sales } = request.body;
     // const { authorization } = request.headers;
     const result = await saleServices.createSale(email, total_price, delivery_address, delivery_number, products_sales);
