+++ back-end/helpers/index.js
@@ -1,9 +1,11 @@
 const STATUS_CODE = require('./statusHelper');
 const STATUS_MESSAGE = require('./msgHelper');
 const { CustomError } = require('./errorHelper');
+const validationsHelper = require('./validationsHelper');
 
 module.exports = {
   STATUS_CODE,
   STATUS_MESSAGE,
   CustomError,
+  validationsHelper,
 };
