+++ back-end/services/userServices.js
@@ -1,54 +1,14 @@
-const { CustomError, STATUS_CODE, STATUS_MESSAGE } = require('../helpers');
+const { CustomError, STATUS_MESSAGE, validationsHelper } = require('../helpers');
 const { generateToken, decodeToken } = require('../auth');
 const { userModels } = require('../models');
 
-const SIX = 6;
-const TWELVE = 12;
-
-const checkIfEmailAndPasswordExist = (email, password) => {
-  if (!email || !password) {
-    throw new CustomError({
-      status: STATUS_CODE.BAD_REQUEST,
-      message: STATUS_MESSAGE.INVALID_ENTIRES,
-    });
-  }
-};
-
-const checkIfEmailIsValid = (email) => {
-  const regex = /\S+@\S+\.\S+/.test(email); // Source: https://stackoverflow.com/questions/46155/how-to-validate-an-email-address-in-javascript
-  if (!regex) {
-    throw new CustomError({
-      status: STATUS_CODE.BAD_REQUEST,
-      message: STATUS_MESSAGE.INVALID_ENTIRES,
-    });
-  }
-};
-
-const checkIfPasswordIsValid = (password) => {
-  if (password.length < SIX) {
-    throw new CustomError({
-      status: STATUS_CODE.UNAUTHORIZED,
-      message: STATUS_MESSAGE.INVALID_ENTIRES,
-    });
-  }
-};
-
-const checksIfUserHasBeenReturned = (result) => {
-  if (!result.length) {
-    throw new CustomError({
-      status: STATUS_CODE.UNPROCESSABLE_ENTITY,
-      message: STATUS_MESSAGE.USER_NOT_FOUND,
-    });
-  }
-};
-
 const userLogin = async (emailFromRequest, password) => {
-  checkIfEmailAndPasswordExist(emailFromRequest, password);
-  checkIfEmailIsValid(emailFromRequest);
-  checkIfPasswordIsValid(password);
+  validationsHelper.checkIfEmailAndPasswordExist(emailFromRequest, password);
+  validationsHelper.checkIfEmailIsValid(emailFromRequest);
+  validationsHelper.checkIfPasswordIsValid(password);
 
   const result = await userModels.findUserByEmail(emailFromRequest);
-  checksIfUserHasBeenReturned(result);
+  validationsHelper.checksIfUserHasBeenReturned(result);
 
   const { name, email, role } = result[0];
   const token = generateToken.create(name, email, role);
@@ -64,47 +24,20 @@ const userLogin = async (emailFromRequest, password) => {
 
 const userEmail = async (email) => {
   const result = await userModels.findUserByEmail(email);
-  checksIfUserHasBeenReturned(result);
+  validationsHelper.checksIfUserHasBeenReturned(result);
 
   return STATUS_MESSAGE.USER_FOUND;
 };
 
-const checkIfNameEmailPasswordAndSellerExist = (name, email, password, seller) => {
-  if (!name || !email || !password || seller === null) {
-    throw new CustomError({
-      status: STATUS_CODE.BAD_REQUEST,
-      message: STATUS_MESSAGE.INVALID_ENTIRES,
-    });
-  }
-};
-
-const checkIfNameIsValid = (name) => {
-  const regex = /[^a-zA-Z ]/g.test(name);
-  if (regex || name.length < TWELVE) {
-    throw new CustomError({
-      status: STATUS_CODE.BAD_REQUEST,
-      message: STATUS_MESSAGE.INVALID_ENTIRES,
-    });
-  }
-};
-
-const checkIfEmailExist = (result) => {
-  if (result.length) {
-    throw new CustomError({
-      status: STATUS_CODE.BAD_REQUEST,
-      message: STATUS_MESSAGE.DUPLICATED_EMAIL,
-    });
-  }
-};
 
 const userRegistration = async (name, email, password, seller) => {
-  checkIfNameEmailPasswordAndSellerExist(name, email, password, seller);
-  checkIfNameIsValid(name);
-  checkIfEmailIsValid(email);
+  validationsHelper.checkIfNameEmailPasswordAndSellerExist(name, email, password, seller);
+  validationsHelper.checkIfNameIsValid(name);
+  validationsHelper.checkIfEmailIsValid(email);
 
   const emailExist = await userModels.findUserByEmail(email);
-  checkIfEmailExist(emailExist);
-  checkIfPasswordIsValid(password);
+  validationsHelper.checkIfEmailExist(emailExist);
+  validationsHelper.checkIfPasswordIsValid(password);
 
   const role = seller ? 'administrator' : 'client'; 
   await userModels.userRegistration(name, email, password, role);
@@ -112,30 +45,12 @@ const userRegistration = async (name, email, password, seller) => {
   return { message: STATUS_MESSAGE.USER_CREATED };
 };
 
-const checkIfEmailAndNameExist = (name, email) => {
-  if (!name || !email) {
-    throw new CustomError({
-      status: STATUS_CODE.BAD_REQUEST,
-      message: STATUS_MESSAGE.INVALID_ENTIRES,
-    });
-  }
-};
-
-const checkIfUserIsOwner = (emailFromBody, emailFromToken) => {
-  if (emailFromBody !== emailFromToken) {
-    throw new CustomError({
-      status: STATUS_CODE.UNAUTHORIZED,
-      message: STATUS_MESSAGE.EMAIL_NOT_EQUAL,
-    });
-  }
-};
-
 const userProfile = async (name, email, authorization) => {
   const decodedToken = decodeToken.decode(authorization);
-  checkIfUserIsOwner(email, decodedToken.email);
-  checkIfEmailAndNameExist(name, email);
-  checkIfNameIsValid(name);
-  checkIfEmailIsValid(email);
+  validationsHelper.checkIfUserIsOwner(email, decodedToken.email);
+  validationsHelper.checkIfEmailAndNameExist(name, email);
+  validationsHelper.checkIfNameIsValid(name);
+  validationsHelper.checkIfEmailIsValid(email);
   await userModels.userProfile(name, email);
   return { message: STATUS_MESSAGE.NAME_UPDATED };
 };
